chore: restucture app to have the router in the router dir
