+++ app/constants.py
@@ -1,0 +1,2 @@
+DEFAULT_CPU = "100m"
+DEFAULT_MEM = "128Mi"
